refactor(models): move build_qwen3_moe_chain into models/qwen3_moe/conversion_chain.py
Consistency: every model's chain now lives in its own package; conversion_chains.py
holds only the shared helpers (_routed_experts_op, _GATE_DOWN_UP) and the registry.
